Fix TOCTOU race in ResolveServer + add missing test assertions
- Rewrite ResolveServer to run inside a transaction with SELECT...FOR
  UPDATE scoped to matching candidates only (by server_uuid OR
  host+port+username), preventing concurrent callers from creating
  duplicate active entries via InnoDB gap locks
- Change Server.Port from uint to uint16 to match TCP port range and
  parseSourceDSN return type; remove now-redundant uint() cast in
  serverutil.go
- Add TestBuildQueryViaGlob: exercises the full buildGlob→buildQuery
  pipeline so the /**/*.parquet glob is verified end-to-end
- Add decommission post-condition assertions to TestServerID_Decommission:
  verify decommissioned_at is set on the old record and the new record
  is persisted with a fresh bintrail_id

+++ cmd/bintrail/serverutil.go
@@ -20,5 +20,5 @@ func resolveServerIdentity(ctx context.Context, sourceDB, indexDB *sql.DB, sourc
 	if err != nil {
 		return "", err
 	}
-	return serverid.ResolveServer(ctx, indexDB, serverUUID, host, uint(port), user)
+	return serverid.ResolveServer(ctx, indexDB, serverUUID, host, port, user)
 }
